ml_service/bot/requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
SQLAlchemy==2.0.25
httpx~=0.25.2
bcrypt==4.1.2
python-dateutil==2.8.2 